make sure the sshKey secret is put into manifestwork
