+++ pkg/controllers/hypershiftdeployment_controller_test.go
@@ -2,7 +2,6 @@ package controllers
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"testing"
@@ -13,16 +12,11 @@ import (
 	"github.com/openshift/hypershift/cmd/infra/aws"
 	"github.com/openshift/hypershift/cmd/infra/azure"
 	hyd "github.com/stolostron/hypershift-deployment-controller/api/v1alpha1"
-	"github.com/stolostron/hypershift-deployment-controller/pkg/helper"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	workv1 "open-cluster-management.io/api/work/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -236,268 +230,6 @@ func genKeyFromObject(obj metav1.Object) types.NamespacedName {
 	return types.NamespacedName{Name: obj.GetName(), Namespace: obj.GetNamespace()}
 }
 
-type kindAndKey struct {
-	schema.GroupVersionKind
-	types.NamespacedName
-}
-
-// TestManifestWorkFlow tests when override is set to manifestwork, test if the manifestwork is created
-// and referenece secret is put into manifestwork payload
-func TestManifestWorkFlow(t *testing.T) {
-	client := initClient()
-	ctx := context.Background()
-
-	infraOut := getAWSInfrastructureOut()
-	testHD := getHypershiftDeployment("default", "test1")
-	testHD.Spec.Override = hyd.InfraConfigureWithManifest
-
-	testHD.Spec.Infrastructure.Platform = &hyd.Platforms{AWS: &hyd.AWSPlatform{}}
-	testHD.Spec.Credentials = &hyd.CredentialARNs{AWS: &hyd.AWSCredentials{}}
-	testHD.Spec.InfraID = infraOut.InfraID
-	ScaffoldAWSHostedClusterSpec(testHD, infraOut)
-	ScaffoldAWSNodePoolSpec(testHD, infraOut)
-
-	client.Create(ctx, testHD)
-	defer client.Delete(ctx, testHD)
-
-	// ensure the pull secret exist in cluster
-	// this pull secret is generated by the hypershift operator
-	pullSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fmt.Sprintf("%s-pull-secret", testHD.GetName()),
-			Namespace: "default",
-		},
-		Data: map[string][]byte{
-			".dockerconfigjson": []byte(`docker-pull-secret`),
-		},
-	}
-
-	client.Create(ctx, pullSecret)
-
-	hdr := &HypershiftDeploymentReconciler{
-		Client: client,
-		Log:    ctrl.Log.WithName("tester"),
-	}
-
-	_, err := hdr.Reconcile(ctx, ctrl.Request{NamespacedName: getNN})
-	assert.Nil(t, err, "err nil when reconcile was successfull")
-
-	manifestWorkKey := types.NamespacedName{
-		Name:      generateManifestName(testHD),
-		Namespace: helper.GetTargetManagedCluster(testHD)}
-
-	manifestWork := &workv1.ManifestWork{}
-
-	err = client.Get(ctx, manifestWorkKey, manifestWork)
-	assert.Nil(t, err, "err nil when manifestwork generated")
-
-	requiredResource := map[kindAndKey]bool{
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "hypershift.openshift.io", Version: "v1alpha1", Kind: "HostedCluster"},
-			NamespacedName: genKeyFromObject(testHD)}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "hypershift.openshift.io", Version: "v1alpha1", Kind: "NodePool"},
-			NamespacedName: genKeyFromObject(testHD)}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "Secret"},
-			NamespacedName: types.NamespacedName{
-				Name: "test1-node-mgmt-creds", Namespace: helper.GetTargetNamespace(testHD)}}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "Secret"},
-			NamespacedName: types.NamespacedName{
-				Name: "test1-cpo-creds", Namespace: helper.GetTargetNamespace(testHD)}}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "Secret"},
-			NamespacedName: types.NamespacedName{
-				Name: "test1-node-mgmt-creds", Namespace: helper.GetTargetNamespace(testHD)}}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "Secret"},
-			NamespacedName: types.NamespacedName{
-				Name: "test1-pull-secret", Namespace: helper.GetTargetNamespace(testHD)}}: false,
-	}
-
-	wl := manifestWork.Spec.Workload.Manifests
-
-	for _, w := range wl {
-		u := &unstructured.Unstructured{}
-		if err := json.Unmarshal(w.Raw, u); err != nil {
-			assert.Nil(t, err, "err nil when convert manifest to unstructured")
-		}
-
-		k := kindAndKey{
-			GroupVersionKind: u.GetObjectKind().GroupVersionKind(),
-			NamespacedName:   genKeyFromObject(u),
-		}
-
-		requiredResource[k] = true
-	}
-
-	for k, v := range requiredResource {
-		assert.True(t, v, fmt.Sprintf("resource %s should be in the manifestWork.Spec.Workload.Manifests", k))
-	}
-}
-
-// TestManifestWorkFlowWithExtraConfigurations test
-// when override is set to manifestwork, test if the manifestwork is created
-// and extra secret/configmap is put into manifestwork payload in addition to
-// the required resource of TestManifestWorkFlow
-func TestManifestWorkFlowWithExtraConfigurations(t *testing.T) {
-	client := initClient()
-	ctx := context.Background()
-
-	infraOut := getAWSInfrastructureOut()
-	testHD := getHypershiftDeployment("default", "test1")
-	testHD.Spec.Override = hyd.InfraConfigureWithManifest
-
-	testHD.Spec.Infrastructure.Platform = &hyd.Platforms{AWS: &hyd.AWSPlatform{}}
-	testHD.Spec.Credentials = &hyd.CredentialARNs{AWS: &hyd.AWSCredentials{}}
-	testHD.Spec.InfraID = infraOut.InfraID
-	ScaffoldAWSHostedClusterSpec(testHD, infraOut)
-	ScaffoldAWSNodePoolSpec(testHD, infraOut)
-
-	cfgSecretName := "hostedcluster-config-secret-1"
-	cfgConfigName := "hostedcluster-config-configmap-1"
-	cfgItemSecretName := "hostedcluster-config-item-1"
-
-	insertConfigSecretAndConfigMap := func() {
-		testHD.Spec.HostedClusterSpec.Configuration = &hyp.ClusterConfiguration{}
-		testHD.Spec.HostedClusterSpec.Configuration.SecretRefs = []corev1.LocalObjectReference{
-			corev1.LocalObjectReference{Name: cfgSecretName}}
-
-		testHD.Spec.HostedClusterSpec.Configuration.ConfigMapRefs = []corev1.LocalObjectReference{
-			corev1.LocalObjectReference{Name: cfgConfigName}}
-
-		testHD.Spec.HostedClusterSpec.Configuration.Items = []runtime.RawExtension{runtime.RawExtension{Object: &corev1.Secret{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Secret",
-				APIVersion: corev1.SchemeGroupVersion.String(),
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      cfgItemSecretName,
-				Namespace: testHD.GetNamespace(),
-			},
-			Data: map[string][]byte{
-				".dockerconfigjson": []byte(`docker-pull-secret`),
-			},
-		},
-		}}
-	}
-
-	insertConfigSecretAndConfigMap()
-
-	// ensure the pull secret exist in cluster
-	// this pull secret is generated by the hypershift operator
-	pullSecret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fmt.Sprintf("%s-pull-secret", testHD.GetName()),
-			Namespace: testHD.GetNamespace(),
-		},
-		Data: map[string][]byte{
-			".dockerconfigjson": []byte(`docker-pull-secret`),
-		},
-	}
-
-	client.Create(ctx, pullSecret)
-	defer client.Delete(ctx, pullSecret)
-
-	se := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cfgSecretName,
-			Namespace: testHD.GetNamespace(),
-		},
-		Data: map[string][]byte{
-			".dockerconfigjson": []byte(`docker-pull-secret`),
-		},
-	}
-
-	client.Create(ctx, se)
-	defer client.Delete(ctx, se)
-
-	cm := &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cfgConfigName,
-			Namespace: testHD.GetNamespace(),
-		},
-		Data: map[string]string{
-			".dockerconfigjson": "docker-configmap",
-		},
-	}
-
-	client.Create(ctx, cm)
-	defer client.Delete(ctx, cm)
-
-	client.Create(ctx, testHD)
-	defer client.Delete(ctx, testHD)
-
-	hdr := &HypershiftDeploymentReconciler{
-		Client: client,
-		Log:    ctrl.Log.WithName("tester"),
-	}
-
-	_, err := hdr.Reconcile(ctx, ctrl.Request{NamespacedName: getNN})
-	assert.Nil(t, err, "err nil when reconcile was successfull")
-
-	manifestWorkKey := types.NamespacedName{
-		Name:      generateManifestName(testHD),
-		Namespace: helper.GetTargetManagedCluster(testHD)}
-
-	manifestWork := &workv1.ManifestWork{}
-
-	err = client.Get(ctx, manifestWorkKey, manifestWork)
-	assert.Nil(t, err, "err nil when manifestwork generated")
-
-	requiredResource := map[kindAndKey]bool{
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "Secret"},
-			NamespacedName: types.NamespacedName{
-				Name: cfgSecretName, Namespace: helper.GetTargetNamespace(testHD)}}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "Secret"},
-			NamespacedName: types.NamespacedName{
-				Name: cfgItemSecretName, Namespace: helper.GetTargetNamespace(testHD)}}: false,
-
-		kindAndKey{
-			GroupVersionKind: schema.GroupVersionKind{
-				Group: "", Version: "v1", Kind: "ConfigMap"},
-			NamespacedName: types.NamespacedName{
-				Name: cfgConfigName, Namespace: helper.GetTargetNamespace(testHD)}}: false,
-	}
-
-	wl := manifestWork.Spec.Workload.Manifests
-
-	for _, w := range wl {
-		u := &unstructured.Unstructured{}
-		if err := json.Unmarshal(w.Raw, u); err != nil {
-			assert.Nil(t, err, "err nil when convert manifest to unstructured")
-		}
-
-		k := kindAndKey{
-			GroupVersionKind: u.GetObjectKind().GroupVersionKind(),
-			NamespacedName:   genKeyFromObject(u),
-		}
-
-		requiredResource[k] = true
-	}
-
-	for k, v := range requiredResource {
-		assert.True(t, v, fmt.Sprintf("resource %s should be in the manifestWork.Spec.Workload.Manifests", k))
-	}
-}
-
 func TestHypershiftdeployment_controller(t *testing.T) {
 
 	client := initClient()
@@ -524,36 +256,4 @@ func TestHypershiftdeployment_controller(t *testing.T) {
 	var testNP hyp.NodePool
 	err = client.Get(context.Background(), getNN, &testNP)
 	assert.Nil(t, err, "err is Nil when NodePool is successfully created")
-}
-
-func TestConfigureFalseWithManifestWork(t *testing.T) {
-
-	client := initClient()
-
-	testHD := getHypershiftDeployment(getNN.Namespace, getNN.Name)
-	testHD.Spec.Override = hyd.InfraConfigureWithManifest
-	testHD.Spec.TargetManagedCluster = "local-host"
-	testHD.Spec.TargetNamespace = "multicluster-engine"
-
-	client.Create(context.Background(), testHD)
-
-	hdr := &HypershiftDeploymentReconciler{
-		Client: client,
-	}
-	_, err := hdr.Reconcile(context.Background(), ctrl.Request{NamespacedName: getNN})
-	assert.Nil(t, err, "err nil when reconcile was successful")
-
-	var resultHD hyd.HypershiftDeployment
-	err = client.Get(context.Background(), getNN, &resultHD)
-	assert.Nil(t, err, "is nil when HypershiftDeployment resource is found")
-	assert.True(t, meta.IsStatusConditionFalse(resultHD.Status.Conditions, string(hyd.WorkConfigured)), "is true when ManifestWork is not configured correctly")
-
-	err = client.Delete(context.Background(), &resultHD)
-	assert.Nil(t, err, "is nill when HypershiftDeployment resource is deleted")
-
-	_, err = hdr.Reconcile(context.Background(), ctrl.Request{NamespacedName: getNN})
-	assert.Nil(t, err, "err nil when reconcile on delete was successful")
-
-	err = client.Get(context.Background(), getNN, &resultHD)
-	assert.True(t, errors.IsNotFound(err), "is not found when HypershiftDeployment resource is deleted successfully")
 }
