+++ pkg/controllers/manifestwork.go
@@ -118,7 +118,6 @@ func (r *HypershiftDeploymentReconciler) createMainfestwork(ctx context.Context,
 	}
 
 	// if the manifestwork is created, then move the status to hypershiftDeployment
-	// TODO: @ianzhang366 might want to do some upate/patch when the manifestwork is created.
 	if err := r.Get(ctx, getManifestWorkKey(hyd), m); err == nil {
 		inHyd := hyd.DeepCopy()
 		syncManifestworkStatusToHypershiftDeployment(hyd, m)
@@ -144,16 +143,18 @@ func (r *HypershiftDeploymentReconciler) createMainfestwork(ctx context.Context,
 		}
 	}
 
-	m.Spec.Workload.Manifests = payload
-
-	// a placeholder for later use
-	noOp := func(in *workv1.ManifestWork, payload []workv1.Manifest) controllerutil.MutateFn {
+	// the in object will get override by a GET
+	// after the GET, the update will be called and the payload will be wrote to
+	// the in object, which will be send with a UPDATE
+	update := func(in *workv1.ManifestWork, payload []workv1.Manifest) controllerutil.MutateFn {
 		return func() error {
+			m.Spec.Workload.Manifests = payload
 			return nil
 		}
 	}
 
-	if _, err := controllerutil.CreateOrUpdate(r.ctx, r.Client, m, noOp(m, payload)); err != nil {
+	_, err = controllerutil.CreateOrUpdate(r.ctx, r.Client, m, update(m, payload))
+	if err != nil {
 		r.Log.Error(err, fmt.Sprintf("failed to CreateOrUpdate the existing manifestwork %s", getManifestWorkKey(hyd)))
 		return ctrl.Result{}, err
 
@@ -211,6 +212,19 @@ func (r *HypershiftDeploymentReconciler) appendHostedClusterReferenceSecrets(ctx
 				return nil
 			}
 			refSecrets = append(refSecrets, ScaffoldAzureCloudCredential(hyd, creds))
+		}
+
+		sshKey := hyd.Spec.HostedClusterSpec.SSHKey
+		if len(sshKey.Name) != 0 {
+			s, err := r.generateSecret(ctx,
+				types.NamespacedName{Name: sshKey.Name,
+					Namespace: hyd.GetNamespace()})
+
+			if err != nil {
+				return fmt.Errorf("failed to duplicateSecret, err %w", err)
+			}
+
+			refSecrets = append(refSecrets, s)
 		}
 
 		for _, s := range refSecrets {
